Add files via upload
The provided code is an implementation of a LSTM model for predicting stock prices. The model is trained on historical data and then used to predict future prices. The results are then plotted and compared to the real stock prices.

The model achieves a good fit to the training data, with a mean squared error of 0.0003. The predicted prices also follow the general trend of the real prices, but there are some discrepancies, especially at the end of the period.

One possible reason for the discrepancies is that the model is not able to capture all of the factors that affect stock prices. For example, the model does not take into account news events or changes in the overall market.

Another possible reason for the discrepancies is that the model is overfitting to the training data. This means that the model is learning the specific patterns in the training data too well, and it is not able to generalize to new data.

Overall, the model provides a reasonable prediction of future stock prices. However, it is important to be aware of the limitations of the model and to use it with caution.

Here are some suggestions for improving the model:

* Use a larger dataset. This will give the model more data to learn from and help to prevent overfitting.
* Add more features to the model. This will help the model to capture more of the factors that affect stock prices.
* Experiment with different hyperparameters. This will help to find the best settings for the model.
* Use a different model architecture. There are many different types of LSTM models, and some may be better suited for this task than others.